requirements.txt
fastapi
uvicorn
pybooru
requests
pandas
pyarrow
aiohttp
pyvips
tqdm
imagehash
Pillow


